User app cleansing
- changed how the user token works
- integrated transactional and token in one
- changed the way a user resets his password
- changes on verifyemail, reset password and create password

+++ quests/contrib/quest_handler.py
@@ -3,12 +3,14 @@
 #All Django Imports
 from django.conf import settings
 from django.http import Http404
-from django.shortcuts import redirect
+from django.shortcuts import redirect, get_object_or_404
 from django.utils import timezone
 from datetime import datetime, timedelta
 
 #All local imports (libs, contribs, models)
 from quests.models import Quests, QuestTransactional, QuestToken, QuestEvents
+from libs.url_shortener import UrlShortener
+
 #All external imports (libs, packages)
 import logging
 import pytz
@@ -22,115 +24,35 @@ def listfeaturedquests(questrs_id):
     allquests = Quests.objects.filter(ishidden=False)
     return allquests
 
-def getQuestsByUser(questrs_id):
-    """List all the quests by a particular user"""
-    try:
-        questsbysuer = Quests.objects.filter(questrs_id=questrs_id, ishidden=False, shipper=None)
-    except Quests.DoesNotExist:
-        raise Http404
+def getQuestsByUser(user):
+    """List all the quests created by a particular user"""
+    questsbysuer = Quests.objects.filter(questrs_id=user.id, ishidden=False)
     return questsbysuer
 
 def getQuestDetails(quest_id):
-    try:
-        questdetails = Quests.objects.get(id=quest_id)
-    except Quests.DoesNotExist:
-        raise Http404
+    questdetails = get_object_or_404(Quests, id=quest_id)
     return questdetails
 
 def getQuestDetailsByTrackingNumber(tracking_number):
-    try:
-        questdetails = Quests.objects.get(tracking_number=tracking_number)
-    except Quests.DoesNotExist:
-        raise Http404
+    questdetails = get_object_or_404(Quests, tracking_number=tracking_number)    
     return questdetails
 
-def getMyShipmnets(shipper_id):
-    myshipments = Quests.objects.filter(shipper=shipper_id)
-    return myshipments
-
-def getQuestsWithOffer(questrs_id):
-    """Lists a user's quest where offers are put"""
-    try:
-        questsWithOffer = Quests.objects.filter(questrs_id=questrs_id, ishidden=False).exclude(shipper=None)
-    except Quests.DoesNotExist:
-        raise Http404
-    return questsWithOffer
-
-def isShipperForQuest(shipper_id, questname):
+def isShipperForQuest(shipper, quest_id):
     """Returns if the current shipper is listed in for the quest, the shipper_id has to be converted to string"""
-    try:
-        questdetails = Quests.objects.get(id=questname)
-    except Quests.DoesNotExist:
-        raise Http404
+    questdetails = getQuestDetails(quest_id)
 
     current_shipper = questdetails.shipper
     if current_shipper != None:
-        if shipper_id in current_shipper:
+        if str(shipper.id) in current_shipper:
             return True
         return False
     return False
-
-def addShipper(shipper_id, questname):
-    """adds a shipper to a posted quest"""
-    try:
-        questdetails = Quests.objects.get(id=questname)
-    except Quests.DoesNotExist:
-        raise Http404
-
-    current_shipper = questdetails.shipper
-    # for first application
-    logger.debug(current_shipper)
-    if current_shipper==None:
-        current_shipper = []
-    else:
-        current_shipper = current_shipper.split(',')
-    # check if shipper has already applied
-    if not shipper_id in current_shipper:
-        current_shipper.append(shipper_id)
-        current_shipper = ','.join(current_shipper)
-    else:
-        return redirect('home')
-
-    try:
-        Quests.objects.filter(id=questname).update(shipper=current_shipper)
-    except Quests.DoesNotExist:
-        raise Http404
-
-def delShipper(shipper_id, questname):
-    """adds a shipper to a posted quest"""
-    try:
-        questdetails = Quests.objects.get(id=questname)
-    except Quests.DoesNotExist:
-        raise Http404
-
-    current_shipper = questdetails.shipper
-    if current_shipper==None:
-        #If no user has bid so far , redirect to home
-        return redirect('home')
-    else:
-        current_shipper = current_shipper.split(',')
-    # check if shipper has already applied
-    if shipper_id in current_shipper:
-        current_shipper.remove(shipper_id)
-        if current_shipper:
-            current_shipper = ','.join(current_shipper)
-        else:
-            current_shipper=None
-    else:
-        return redirect('home')
-
-    try:
-        Quests.objects.filter(id=questname).update(shipper=current_shipper)
-    except Quests.DoesNotExist:
-        raise Http404
 
 def prepNewQuestNotification(user, questdetails, accept_url, reject_url):
     """Prepare the details for notification emails for new quests"""
     template_name="New_Quest_Notification"
     subject="New Quest Notification"
-    # quest_browse_link=settings.QUESTR_URL+"/quest"
     quest_support_email="support@questr.co"
-    questr_unsubscription_link="http://questr.co/unsub"
 
     email_details = {
                         'subject' : subject,
@@ -145,7 +67,6 @@ def prepNewQuestNotification(user, questdetails, accept_url, reject_url):
                                                 'quest_distance'      : str(questdetails.distance),
                                                 'quest_creation_date'      : questdetails.creation_date.strftime('%m-%d-%Y'),
                                                 'quest_support_mail': quest_support_email,
-                                                'questr_unsubscription_link' : questr_unsubscription_link,
                                                 'company'           : "Questr Co",
                                                 'quest_accept_url'           : accept_url,
                                                 'quest_reject_url'           : reject_url,
@@ -161,7 +82,6 @@ def prepNewQuestAdminNotification(user, questdetails):
     """Prepare the details for notification emails for new quests to admins"""
     template_name="New_Quest_Admin_Notification"
     subject="New Quest Notification"
-    # quest_browse_link=settings.QUESTR_URL+"/quest"
     quest_support_email="support@questr.co"
 
     email_details = {
@@ -196,7 +116,6 @@ def prepOfferAcceptedNotification(user, questdetails):
     """Prepare the details for notification emails for new quests to admins"""
     template_name="Offer_Accepted_Notification"
     subject="Questr - Offer Accepted"
-    # quest_browse_link=settings.QUESTR_URL+"/quest"
     quest_support_email="support@questr.co"
 
     email_details = {
@@ -231,9 +150,7 @@ def prepQuestAppliedNotification(shipper, questr, questdetails):
     """Prepare the details for notification emails for new quests"""
     template_name="Quest_Accepted_Notification_Questr"
     subject="Questr - Your shipment has been processed"
-    # quest_browse_link=settings.QUESTR_URL+"/quest"
     quest_support_email="support@questr.co"
-    questr_unsubscription_link="http://questr.co/unsub"
 
     email_details = {
                         'subject' : subject,
@@ -247,7 +164,6 @@ def prepQuestAppliedNotification(shipper, questr, questdetails):
                                                 'shipper_user_name': shipper.displayname,                                                
                                                 'shipper_phone': shipper.phone,                                                
                                                 'shipper_profile_link'  : settings.QUESTR_URL+'/user/'+shipper.displayname,                                                
-                                                'email_unsub_link'  : questr_unsubscription_link,
                                                 'quest_title'       : questdetails.title,
                                                 'quest_size'      : questdetails.size,
                                                 'quest_pickup_name' : questdetails.pickup['name'],
@@ -276,9 +192,7 @@ def prepQuestCompleteNotification(shipper, questr, questdetails, review_link):
     """Prepare the details for notification emails for completion of quests"""
     template_name="Quest_Completion_Notification"
     subject="Questr - Quest Completed"
-    # quest_browse_link=settings.QUESTR_URL+"/quest"
     quest_support_email="support@questr.co"
-    questr_unsubscription_link="http://questr.co/unsub"
 
     # Calculating the delivery time for the quest
     quest_creation_date = questdetails.creation_date
@@ -304,7 +218,6 @@ def prepQuestCompleteNotification(shipper, questr, questdetails, review_link):
                                                 'quest_creation_date'      : questdetails.creation_date.strftime('%m-%d-%Y'),
                                                 'delivery_time'      : delivery_time,
                                                 'quest_support_mail': quest_support_email,
-                                                'questr_unsubscription_link' : questr_unsubscription_link,
                                                 'company'           : "Questr Co"
 
                                                 },
@@ -327,19 +240,18 @@ def update_resized_image(quest_id):
     logger.debug(file_path)
     filename_base, filename_ext = os.path.splitext(file_path)
     normal_file_path = "%s_%s_normal.jpg" % (filename_base, questdetails.id)
-    logger.debug(normal_file_path)
-    try:
-        Quests.objects.filter(id=quest_id).update(item_images=normal_file_path)
-    except Quests.DoesNotExist:
-        raise Http404
-    logger.debug(storage.exists(file_path))
+    logger.debug("Normal file path for {0} is {1}".format(quest_id, normal_file_path))
+    queryset = Quests.objects.filter(id=quest_id).update(item_images=normal_file_path)
+    if queryset == 0:
+        logger.debug("Image for {0} couldn't be updated as {0} doesn't exist".format(quest_id))
+        return ""
+    logger.debug("Image for {0} has been updated, found at is {1}".format(quest_id, storage.url(normal_file_path)))    
     if storage.exists(file_path):
         storage.delete(file_path)
     return ""
 
 def updateQuestWithAvailableCourierDetails(quest, available_shippers):
     """Updates the available_shippers field in the given quest with the available couriers and their details"""
-    logger.debug("trying to update available couriers")
     Quests.objects.filter(id=quest.id).update(available_couriers=available_shippers)
     if quest.considered_couriers == '[]':
         considered_couriers = [int(x) for x in available_shippers]
@@ -365,7 +277,6 @@ def get_accept_url(quest=None, shipper=None):
         questr_token.save()
         accept_link = "{0}/quest/accept/{1}?quest_token={2}".format(settings.QUESTR_URL , transcational.get_truncated_quest_code(), token_id)
     
-    from libs.url_shortener import UrlShortener
     shortener = UrlShortener()
     short_url = shortener.get_short_url(dict(longUrl=accept_link))
     logger.warn("short url for accepting the quest {0} for courier {1} is {2}".format(quest, shipper, short_url))
@@ -394,7 +305,6 @@ def get_reject_url(quest=None, shipper=None):
         questr_token.save()
         reject_link = "{0}/quest/reject/{1}?quest_token={2}".format(settings.QUESTR_URL , transcational.get_truncated_quest_code(), token_id)
 
-    from libs.url_shortener import UrlShortener
     shortener = UrlShortener()
     short_url = shortener.get_short_url(dict(longUrl=reject_link))
     logger.warn("short url for rejecting the quest {0} for courier {1} is {2}".format(quest, shipper, short_url))
